+++ cli/utils/user-config.js
@@ -2,6 +2,7 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 const _ = require("./frequent");
 const path = require("path");
+const macros_1 = require("../../lib/lexer/grammar/macros");
 const fthtmlconfig = path.resolve(process.cwd(), 'fthtmlconfig.json');
 const config = {
     rootDir: null,
@@ -9,7 +10,9 @@ const config = {
     excluded: [],
     importDir: null,
     exportDir: null,
-    templateDir: null
+    jsonDir: null,
+    prettify: false,
+    globalvars: {}
 };
 if (_.fileExists(fthtmlconfig)) {
     const parsed = _.getJSONFromFile(fthtmlconfig);
@@ -17,15 +20,25 @@ if (_.fileExists(fthtmlconfig)) {
         config.rootDir = path.resolve(parsed.rootDir);
     if (_.isTypeOf(parsed.keepTreeStructure, 'boolean'))
         config.keepTreeStructure = parsed.keepTreeStructure;
+    if (_.isTypeOf(parsed.prettify, 'boolean'))
+        config.prettify = parsed.prettify;
     if (Array.isArray(parsed.excluded)) {
         if (_.isTypesOfAndNotEmpty(parsed.excluded, 'string'))
             config.excluded = parsed.excluded;
     }
     if (_.isTypeOfAndNotEmpty(parsed.importDir, 'string'))
         config.importDir = path.resolve(parsed.importDir);
+    if (_.isTypeOfAndNotEmpty(parsed.jsonDir, 'string'))
+        config.jsonDir = path.resolve(parsed.jsonDir);
     if (_.isTypeOfAndNotEmpty(parsed.exportDir, 'string'))
         config.exportDir = path.resolve(parsed.exportDir);
-    if (_.isTypeOfAndNotEmpty(parsed.templateDir, 'string'))
-        config.templateDir = path.resolve(parsed.templateDir);
+    if (_.isTypeOf(parsed.globalvars, 'object')) {
+        if (_.isTypesOfAndNotEmpty(parsed.globalvars, 'string'))
+            config.globalvars = parsed.globalvars;
+        Object.keys(config.globalvars).forEach(gvar => {
+            if (macros_1.default[gvar] !== undefined)
+                delete config.globalvars[gvar];
+        });
+    }
 }
 exports.default = config;