+++ cli/utils/user-config.ts
@@ -1,5 +1,6 @@
 import * as _ from "./frequent";
 import * as path from "path";
+import { default as macros } from "../../lib/lexer/grammar/macros";
 
 const fthtmlconfig = path.resolve(process.cwd(), 'fthtmlconfig.json');
 
@@ -9,7 +10,9 @@ const config = {
     excluded: <string[]>[],
     importDir: <string>null,
     exportDir: <string>null,
-    templateDir: <string>null
+    jsonDir: <string>null,
+    prettify: <boolean>false,
+    globalvars: {}
 }
 
 if (_.fileExists(fthtmlconfig)) {
@@ -23,6 +26,10 @@ if (_.fileExists(fthtmlconfig)) {
     if (_.isTypeOf(parsed.keepTreeStructure, 'boolean'))
         config.keepTreeStructure = parsed.keepTreeStructure;
 
+    //prettify
+    if (_.isTypeOf(parsed.prettify, 'boolean'))
+        config.prettify = parsed.prettify;
+
     //excluded
     if (Array.isArray(parsed.excluded)) {
         if (_.isTypesOfAndNotEmpty(parsed.excluded, 'string'))
@@ -33,13 +40,24 @@ if (_.fileExists(fthtmlconfig)) {
     if (_.isTypeOfAndNotEmpty(parsed.importDir, 'string'))
         config.importDir = path.resolve(parsed.importDir);
 
+    //jsonDir
+    if (_.isTypeOfAndNotEmpty(parsed.jsonDir, 'string'))
+        config.jsonDir = path.resolve(parsed.jsonDir);
+
     //exportDir
     if (_.isTypeOfAndNotEmpty(parsed.exportDir, 'string'))
         config.exportDir = path.resolve(parsed.exportDir);
 
-    //templateDir
-    if (_.isTypeOfAndNotEmpty(parsed.templateDir, 'string'))
-        config.templateDir = path.resolve(parsed.templateDir);
+    //global vars
+    if (_.isTypeOf(parsed.globalvars, 'object')) {
+        if (_.isTypesOfAndNotEmpty(parsed.globalvars, 'string'))
+            config.globalvars = parsed.globalvars;
+
+        Object.keys(config.globalvars).forEach(gvar => {
+            if (macros[gvar] !== undefined)
+                delete config.globalvars[gvar]
+        })
+    }
 }
 
 export default config;