+++ lib/lexer/types.ts
@@ -1,4 +1,4 @@
-import { token } from "./token";
+import { token, TOKEN_TYPE as TT } from "./token";
 
 export type char = string;
 export type Tokenable = token | null;
@@ -6,6 +6,18 @@ export type TokenStream = {
     next: () => Tokenable;
     peek: () => Tokenable;
     eof: () => boolean;
+};
+
+export type funcrules = {
+    argsSequenceStrict: boolean;
+    argPatterns: {
+        type: TT[],
+        name: string,
+        possibleValues?: string[],
+        isOptional?: boolean,
+        isRestParameter?: boolean
+    }[];
+    do: (...values: any[]) => any;
 }
 
 export { LEX_MODE } from './lexmode';